requirements.txt
# Core ML libraries
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration and logging
pyyaml>=6.0
omegaconf>=2.3.0
matplotlib>=3.7.0
plotly>=5.15.0
tensorboard>=2.13.0

# Tabular/EHR specific
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep learning for tabular data
pytorch-tabnet>=4.0
rtdl>=0.0.8

# Explainability
shap>=0.42.0
lime>=0.2.0

# Evaluation and calibration
lifelines>=0.27.0
scikit-plot>=0.3.7

# Visualization and demo
streamlit>=1.25.0
gradio>=3.40.0

# Development and testing
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Advanced features
wandb>=0.15.0
mlflow>=2.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
